Do not perform image upsizing
In a project working with a colleague, we found that when images are smaller than the default size (1024), e.g., `250 x 250`, the runtime was longer than expected because of the upsizing process. However, it appeared to us that the upsizing would not help in detection accuracy. We ended up monkey-patched this proposed change to `resize_image`, and saw significant improvement to runtime, without sacrificing accuracy. We suggest that the upsizing is probably not necessary (or perhaps making it optional).

I also suggest fixing the order: `img_h, img_w`, as the first axis of a NumPy image array gives the image height. Even though this does not affect the result in this function, in the future some contributors may get the wrong idea.